# Conflicts:
#	python/setup.py

+++ python/boomer/boosting/cpp/linalg.h
@@ -19,10 +19,10 @@ namespace linalg {
      * @param n The number of elements in the array `a`
      * @return  A scalar of type `float64`, representing the square of the L2 norm of the given array
      */
-    static inline float64 l2NormPow(const float64* a, intp n) {
+    static inline float64 l2NormPow(const float64* a, uint32 n) {
         float64 result = 0;
 
-        for (intp i = 0; i < n; i++) {
+        for (uint32 i = 0; i < n; i++) {
             result += pow(a[i], 2);
         }
 
@@ -32,10 +32,10 @@ namespace linalg {
     /**
      * Calculates and returns the n-th triangular number, i.e., the number of elements in a n times n triangle.
      *
-     * @param n A scalar of type `intp`, representing the order of the triangular number
-     * @return  A scalar of type `intp`, representing the n-th triangular number
+     * @param n A scalar of type `uint32`, representing the order of the triangular number
+     * @return  A scalar of type `uint32`, representing the n-th triangular number
      */
-    static inline intp triangularNumber(intp n) {
+    static inline uint32 triangularNumber(uint32 n) {
         return (n * (n + 1)) / 2;
     }
